+++ clients/admin-ui/src/app/store.ts
@@ -4,6 +4,10 @@ import { createWrapper } from "next-redux-wrapper";
 
 import { datasetApi, reducer as datasetReducer } from "~/features/dataset";
 import { systemApi } from "~/features/system";
+import {
+  dataCategoriesApi,
+  reducer as dataCategoriesReducer,
+} from "~/features/taxonomy";
 import { reducer as userReducer } from "~/features/user";
 
 const makeStore = () => {
@@ -11,13 +15,16 @@ const makeStore = () => {
     reducer: {
       user: userReducer,
       dataset: datasetReducer,
+      dataCategories: dataCategoriesReducer,
       [datasetApi.reducerPath]: datasetApi.reducer,
       [systemApi.reducerPath]: systemApi.reducer,
+      [dataCategoriesApi.reducerPath]: dataCategoriesApi.reducer,
     },
     middleware: (getDefaultMiddleware) =>
       getDefaultMiddleware().concat(
         datasetApi.middleware,
-        systemApi.middleware
+        systemApi.middleware,
+        dataCategoriesApi.middleware
       ),
   });
   setupListeners(store.dispatch);